Uses weaker permissions in welcome screen
The welcome screen currently requests permission for read/write access to Drive, in order to display the "open" dialog box.  However, all that is needed to display this dialog is readonly access to metadata.  This commit changes to using this weaker scope.  One feature that does require full drive access is the ability to upload files from the "open" dialog, so this feature is disabled.