Changes client ID for Google Drive access
Changes the client ID used for load/save and realtime access.  This has no affect on users as Google Drive does not are which client ID is used to access it.  Only the Cloud Console project matters, and the new client ID belongs to the same cloud console project as the old on.