Fixes broken user icon
The code that ensures the uri scheme/protocol is set to https, was broken as it added https: to the uri string even if the uri already had a scheme/protocol set.  This code uses a new method where the uri is parsed, and its scheme/protocol is overriden to https.