Refactors kernel.py and adds raw_input support
Refactors kernel.py, replacing the main loop, which blocks on the pepper message queue, with a loop that deals pepper messages to various queues.  The main loop is run on a separate thread, and blocks on the shell input queue.

This new architecture allows us to block on specific streams, in particular it allows _raw_input() to block on the stdin queue.

This design will also allow future implementations to implement interrupting the kernel.